pyproject.toml
[project]
name = "claude-code-llm-router"
version = "5.2.0"
description = "Multi-LLM router MCP server for Claude Code — smart complexity routing, Claude subscription monitoring, Codex integration, 20+ providers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<3.14"
authors = [
    { name = "ypollak2", email = "ypollak2@users.noreply.github.com" },
]
keywords = ["llm-router", "llm", "mcp", "claude", "router", "ai", "openai", "gemini", "litellm", "claude-code", "claude-desktop", "cost-optimization", "model-routing", "mcp-server"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.0.0",
    "litellm>=1.50.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiosqlite>=0.20.0",
    "aiohttp>=3.9.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.3",
    "ruff>=0.4",
]
scripts = [
    "httpx>=0.27",
    "pyyaml>=6.0",
]
agno = [
    "agno>=2.5.14",
]
tracing = [
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
]

[project.urls]
Homepage = "https://github.com/ypollak2/llm-router"
Repository = "https://github.com/ypollak2/llm-router"
Issues = "https://github.com/ypollak2/llm-router/issues"
Changelog = "https://github.com/ypollak2/llm-router/blob/main/CHANGELOG.md"

[project.scripts]
llm-router = "llm_router.cli:main"
llm-router-sse = "llm_router.server:main_sse"
llm-router-onboard = "llm_router.onboard:main"
llm-router-install-hooks = "llm_router.install_hooks:main"
llm-router-quickstart = "llm_router.quickstart:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_router"]

[tool.hatch.build.targets.sdist]
exclude = [
    ".claude/",
    ".claude-plugin/",
    ".serena/",
    ".playwright-mcp/",
    ".github/",
    "agents/",
    "demo/",
    "docs/",
    "tests/",
    "*.png",
    "*.jpg",
    "dashboard.png",
    "stitch*.png",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
timeout = 30
filterwarnings = [
    # aiosqlite's worker thread may attempt call_soon_threadsafe after the
    # session-scoped event loop is torn down by pytest-asyncio.
    "ignore:Exception in thread.*_connection_worker_thread:pytest.PytestUnhandledThreadExceptionWarning",
]

[tool.ruff]
line-length = 100
target-version = "py310"
